+++ packages/trpc-server/src/index.test.ts
@@ -1,6 +1,6 @@
 import { initTRPC } from '@trpc/server'
 import { Hono } from 'hono'
-import { z } from 'zod'
+import * as z from 'zod'
 import { trpcServer } from '.'
 
 describe('tRPC Adapter Middleware', () => {