+++ packages/zod-validator/src/v3.test.ts
@@ -1,12 +1,10 @@
 import { Hono } from 'hono'
+import type { ExtractSchema } from 'hono/types'
 import type { ContentfulStatusCode } from 'hono/utils/http-status'
 import type { Equal, Expect } from 'hono/utils/types'
 import { vi } from 'vitest'
 import { z } from 'zod/v3'
 import { zValidator } from '.'
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-type ExtractSchema<T> = T extends Hono<infer _, infer S> ? S : never
 
 describe('Basic', () => {
   const app = new Hono()
@@ -55,7 +53,7 @@ describe('Basic', () => {
             | undefined
         }
         output: {
-          success: boolean
+          success: true
           message: string
           queryName: string | undefined
         }