+++ packages/ua-blocker/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hono/ua-blocker
 
+## 0.1.25
+
+### Patch Changes
+
+- [#1788](https://github.com/honojs/middleware/pull/1788) [`90215c11336aa61772741c46642b3c3f5d302d3e`](https://github.com/honojs/middleware/commit/90215c11336aa61772741c46642b3c3f5d302d3e) Thanks [@github-actions](https://github.com/apps/github-actions)! - chore(ua-blocker): sync `robots.json` with upstream
+
+## 0.1.24
+
+### Patch Changes
+
+- [#1744](https://github.com/honojs/middleware/pull/1744) [`59d89520df19dc9d5bf8a725b8ed584ea1f78e76`](https://github.com/honojs/middleware/commit/59d89520df19dc9d5bf8a725b8ed584ea1f78e76) Thanks [@github-actions](https://github.com/apps/github-actions)! - chore(ua-blocker): sync `robots.json` with upstream
+
 ## 0.1.23
 
 ### Patch Changes